Extract POSHeaderTitleView
