+++ WooCommerce/Classes/POS/Presentation/ItemListView.swift
@@ -40,7 +40,7 @@ private extension ItemListView {
     var headerView: some View {
         VStack {
             HStack {
-                headerTextView
+                POSHeaderTitleView()
                 if !viewModel.shouldShowHeaderBanner {
                     Spacer()
                     Button(action: {
@@ -106,14 +106,6 @@ private extension ItemListView {
         }
     }
 
-    var headerTextView: some View {
-        Text(Localization.productSelectorTitle)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .padding(.vertical, Constants.headerPadding)
-            .font(Constants.titleFont)
-            .foregroundColor(Color.posPrimaryTexti3)
-    }
-
     @ViewBuilder
     func listView(_ items: [POSItem]) -> some View {
         ScrollView {
@@ -135,7 +127,6 @@ private extension ItemListView {
 ///
 private extension ItemListView {
     enum Constants {
-        static let titleFont: Font = .system(size: 40, weight: .bold, design: .default)
         static let bannerTitleFont: Font = .system(size: 24, weight: .bold, design: .default)
         static let bannerSubtitleFont: Font = .system(size: 16, weight: .medium, design: .default)
         static let bannerHeight: CGFloat = 164
@@ -146,16 +137,10 @@ private extension ItemListView {
         static let bannerInfoIconSize: CGFloat = 44
         static let closeIconSize: CGFloat = 26
         static let iconPadding: CGFloat = 26
-        static let headerPadding: CGFloat = 8
         static let itemListPadding: CGFloat = 16
     }
 
     enum Localization {
-        static let productSelectorTitle = NSLocalizedString(
-            "pos.itemlistview.productSelectorTitle",
-            value: "Products",
-            comment: "Title of the Point of Sale product selector"
-        )
         static let headerBannerTitle = NSLocalizedString(
             "pos.itemlistview.headerBannerTitle",
             value: "Showing simple products only",
