+++ WooCommerce/Classes/POS/Utils/Font+WooCommercePOS.swift
@@ -6,4 +6,7 @@ extension Font {
     static let posDetail: Font = Font.system(size: UIFontMetrics.default.scaledValue(for: 16))
     static let posModalBody: Font = Font.system(size: UIFontMetrics.default.scaledValue(for: 24))
     static let posModalTitle: Font = Font.system(size: UIFontMetrics.default.scaledValue(for: 36), weight: .bold)
+    static let posHeaderTitle: Font = Font.system(size: 40,
+                                                  weight: .bold,
+                                                  design: .default)
 }
